fixes for raspberry pi and 32bit os

+++ source/smarties/Core/Launcher.cpp
@@ -160,8 +160,8 @@ void Launcher::createGoRunDir(char* initDir, Uint folderID, MPI_Comm envAppCom)
 
   while(true)
   {
-    snprintf(newDir, 1024, "%s/%s_%03lu_%05lu",
-        initDir, "simulation", folderID, iter);
+    snprintf(newDir, 1024, "%s/%s_%03u_%05u",
+        initDir, "simulation", (unsigned) folderID, (unsigned) iter);
     if ( stat(newDir, &fileStat) >= 0 ) iter++; // directory already exists
     else
     {
